Move create_boilerplate out of astra.cli so scaffolding imports fast
`create_boilerplate` is astra's public scaffold API (#99), meant for
downstream tools to scaffold into their own projects. But reaching it
means `from astra.cli import create_boilerplate`, and importing
`astra.cli` costs ~0.5 s: Click, Rich, and the validation stack, which
pulls `astra.datamodel` and through it `linkml_runtime` — roughly 70% of
the total. Writing two template files should not cost that.

Split into stdlib-only modules:

- `astra/scaffold.py` — `create_boilerplate` and its helper, moved
  verbatim. Imports 8x faster (~0.49 s -> ~0.06 s) and loads neither
  linkml_runtime, pydantic, nor click.
- `astra/spec_version.py` — `installed_spec_version()`, four lines of
  `importlib.metadata` that the scaffold needs and that previously sat in
  `validation/schema.py`, one import away from the whole datamodel.

Nothing is removed from the public surface: `astra.cli` re-exports
`create_boilerplate`, and `astra.validation` still exports
`installed_spec_version`. `astra init` is unaffected — it goes through
the same function.

Two new tests: one pins the import contract in a subprocess (scaffolding
must not pull a heavy dependency), one pins the `astra.cli` re-export so
downstream callers using the old path keep working.
