+++ src/astra/cli.py
@@ -30,13 +30,15 @@ from astra.helpers import (
     load_yaml,
     save_yaml,
 )
+from astra.scaffold import create_boilerplate
 from astra.validation.schema import (
     check_spec_version,
-    installed_spec_version,
     validate_analysis_data,
     validate_universe_data,
 )
 from astra.validation.semantic import validate_analysis, validate_universe_file
+
+__all__ = ["create_boilerplate", "main"]
 
 console = Console()
 
@@ -189,86 +191,6 @@ __pycache__/
 
     if check_only and not converged:
         raise SystemExit(1)
-
-
-def create_boilerplate(directory: Path) -> None:
-    """Write the boilerplate spec scaffold into ``directory``.
-
-    Creates ``universes/`` and writes the boilerplate ``astra.yaml``
-    and ``universes/baseline.yaml``. Touches nothing
-    else — no ``.gitignore``, no git init, no emptiness checks — so
-    downstream tools (e.g. lightcone-cli) can scaffold into existing
-    directories under their own conventions. Existing files are
-    overwritten; callers guard on ``astra.yaml`` presence.
-    """
-    directory.mkdir(parents=True, exist_ok=True)
-    (directory / "universes").mkdir(parents=True, exist_ok=True)
-    _create_boilerplate_astra_yaml(directory)
-
-
-def _create_boilerplate_astra_yaml(directory: Path) -> None:
-    """Create boilerplate astra.yaml with TODOs."""
-    name = directory.name if directory != Path(".") else "My Analysis"
-    spec_version = installed_spec_version() or "1.0"
-
-    astra_yaml = f"""# ASTRA Analysis Specification
-
-version: "{spec_version}"
-name: "{name}"
-container: python:3.12-slim
-description: |
-  TODO: One-paragraph overview of the analysis — its question,
-  scope, and what the reader should take away. A richer write-up
-  (figures, citations, multi-page structure) is authored separately
-  as a report that references this analysis's elements; see the
-  ASTRA documentation.
-
-inputs:
-  - id: primary_data
-    type: data
-    description: "TODO: Describe your primary data source"
-
-outputs:
-  - id: main_result
-    type: metric
-    description: "TODO: Describe your primary output metric"
-    decisions: [example_method]
-    recipe:
-      command: python src/main.py --method {{decisions.example_method}} --out {{output}}
-
-  - id: conclusion
-    type: report
-    description: "Summary of analysis findings"
-    inputs: [main_result]
-    recipe:
-      command: python src/main.py --result {{inputs.main_result}} --out {{output}}
-
-decisions:
-  example_method:
-    label: "Example Method Choice"
-    rationale: "TODO: Explain why this decision matters"
-    default: option_a
-    options:
-      option_a:
-        label: "Option A"
-        description: "TODO: Describe option A"
-      option_b:
-        label: "Option B"
-        description: "TODO: Describe option B"
-"""
-    (directory / "astra.yaml").write_text(astra_yaml)
-
-    # Create baseline universe
-    baseline_universe = """# Baseline Universe
-# Default configuration using standard practices
-
-id: baseline
-description: "Default configuration using standard practices"
-
-decisions:
-  example_method: option_a
-"""
-    (directory / "universes" / "baseline.yaml").write_text(baseline_universe)
 
 
 def _init_git_repo(directory: Path, no_git: bool, quiet: bool = False) -> None:
